looking good!
type in a ticker - i will find the strike price that is closest to the last quote of the ticker and compute volatility. using this volatiliy we simulate various trajectories for the stock.

next, we need to compute the standard deviate for each simulated date and then plot that with the historical. almost there!!